backend_SaaS/requirements.txt
openai>=1.0
google-search-results>=2.4.2
python-dotenv>=1.0
requests>=2.0
